+++ src/broadcastServer.ts
@@ -22,6 +22,9 @@
  * limitations under the License.
  */
 
+// eslint-disable-next-line no-console
+if (process.argv.includes('--loader') || process.argv.includes('-loader')) console.log('\u001B[32mBroadcastServer loaded.\u001B[40;0m');
+
 import { EventEmitter } from 'node:events';
 import { BroadcastChannel } from 'node:worker_threads';